# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Farm2Fork marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /marketplace/listings/{id}:
    get:
      operationId: getMarketplaceListing
      tags: [marketplace]
      summary: Get an active database crop listing
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Active crop listing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FarmerListing" }
        "404":
          description: Listing not found
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List the authenticated buyer's orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
        "401": { description: Authentication required }
        "403": { description: Buyer role required }
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order for an active listing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderInput" }
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { description: Authentication required }
        "403": { description: Buyer role required }
        "409": { description: Listing unavailable or insufficient quantity }
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an authorized order
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { description: Authentication required }
        "404": { description: Order not found }
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel a pending buyer order
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Cancelled order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { description: Authentication required }
        "404": { description: Order not found }
        "409": { description: Order cannot be cancelled }
...
  /farmer/orders:
    get:
      operationId: listFarmerOrders
      tags: [farmer-orders]
      summary: List orders containing the authenticated farmer's listings
      responses:
        "200":
          description: Farmer orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
        "401": { description: Authentication required }
        "403": { description: Farmer role required }
  /farmer/orders/{id}/status:
    patch:
      operationId: updateFarmerOrderStatus
      tags: [farmer-orders]
      summary: Advance an order status as its farmer
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderStatusInput" }
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { description: Authentication required }
        "403": { description: Farmer role required or unauthorized order }
        "409": { description: Invalid status transition }
...
components:
  schemas:
...
    FarmerListing:
      type: object
      properties:
        id: { type: string }
        farmer_user_id: { type: [string, "null"] }
        farmer_name: { type: [string, "null"] }
        farmer_photo_url: { type: [string, "null"] }
        crop_name: { type: string }
        variety: { type: [string, "null"] }
        quantity: { type: number }
        unit: { type: string }
        price: { type: number }
        quality: { type: [string, "null"] }
        location: { type: string }
        harvest_date: { type: [string, "null"], format: date }
        image_url: { type: [string, "null"] }
        status: { type: string, enum: [ACTIVE, SOLD, INACTIVE] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, crop_name, variety, quantity, unit, price, quality, location, harvest_date, image_url, status, created_at, updated_at]
...
    Order:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [PENDING, ACCEPTED, REJECTED, PROCESSING, OUT_FOR_DELIVERY, DELIVERED, CANCELLED] }
        total_amount: { type: number }
        delivery_location: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        buyer_name: { type: [string, "null"] }
        farmer_name: { type: [string, "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
      required: [id, status, total_amount, delivery_location, created_at, updated_at, items]
    OrderItem:
      type: object
      properties:
        id: { type: string }
        crop_listing_id: { type: string }
        crop_name: { type: string }
        variety: { type: [string, "null"] }
        image_url: { type: [string, "null"] }
        quantity: { type: number }
        unit: { type: string }
        price: { type: number }
        subtotal: { type: number }
      required: [id, crop_listing_id, crop_name, variety, image_url, quantity, unit, price, subtotal]
    OrderInput:
      type: object
      properties:
        crop_listing_id: { type: string }
        quantity: { type: number, exclusiveMinimum: 0 }
        delivery_location: { type: string, minLength: 1, maxLength: 240 }
      required: [crop_listing_id, quantity, delivery_location]
    OrderStatusInput:
      type: object
      properties:
        status: { type: string, enum: [PENDING, ACCEPTED, REJECTED, PROCESSING, OUT_FOR_DELIVERY, DELIVERED, CANCELLED] }
      required: [status]
...